use single conftest, move test_system test to test_introspection.py
